pyproject.toml
[project]
name = "mt-saas-api"
version = "0.1.0"
description = "multi-tenant saas api (rbac + billing + webhooks)"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "pydantic-settings>=2.5.2",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
  "redis>=5.2.0",
  "alembic>=1.13.3",
  "PyJWT>=2.9.0",
  "email-validator>=2.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
